+++ py-polars/src/functions/lazy.rs
@@ -1,11 +1,17 @@
 use polars::lazy::dsl;
 use polars::lazy::dsl::Expr;
+use polars::prelude::*;
 use pyo3::prelude::*;
 
 use crate::conversion::Wrap;
-use crate::lazy::ToExprs;
-use crate::prelude::{vec_extract_wrapped, DataType};
+use crate::lazy::{binary_lambda, ToExprs};
+use crate::prelude::{vec_extract_wrapped, DataType, DatetimeArgs, DurationArgs};
 use crate::{PyDataFrame, PyExpr, PyLazyFrame, PyPolarsErr};
+macro_rules! set_unwrapped_or_0 {
+    ($($var:ident),+ $(,)?) => {
+        $(let $var = $var.map(|e| e.inner).unwrap_or(dsl::lit(0));)+
+    };
+}
 
 #[pyfunction]
 pub fn arange(start: PyExpr, end: PyExpr, step: i64) -> PyExpr {
@@ -88,7 +94,86 @@ pub fn cov(a: PyExpr, b: PyExpr) -> PyExpr {
 }
 
 #[pyfunction]
+pub fn cumfold(acc: PyExpr, lambda: PyObject, exprs: Vec<PyExpr>, include_init: bool) -> PyExpr {
+    let exprs = exprs.to_exprs();
+
+    let func = move |a: Series, b: Series| binary_lambda(&lambda, a, b);
+    dsl::cumfold_exprs(acc.inner, func, exprs, include_init).into()
+}
+
+#[pyfunction]
+pub fn cumreduce(lambda: PyObject, exprs: Vec<PyExpr>) -> PyExpr {
+    let exprs = exprs.to_exprs();
+
+    let func = move |a: Series, b: Series| binary_lambda(&lambda, a, b);
+    dsl::cumreduce_exprs(func, exprs).into()
+}
+
+#[pyfunction]
+pub fn datetime(
+    year: PyExpr,
+    month: PyExpr,
+    day: PyExpr,
+    hour: Option<PyExpr>,
+    minute: Option<PyExpr>,
+    second: Option<PyExpr>,
+    microsecond: Option<PyExpr>,
+) -> PyExpr {
+    let year = year.inner;
+    let month = month.inner;
+    let day = day.inner;
+
+    set_unwrapped_or_0!(hour, minute, second, microsecond);
+
+    let args = DatetimeArgs {
+        year,
+        month,
+        day,
+        hour,
+        minute,
+        second,
+        microsecond,
+    };
+    dsl::datetime(args).into()
+}
+
+#[pyfunction]
 pub fn dtype_cols(dtypes: Vec<Wrap<DataType>>) -> PyResult<PyExpr> {
     let dtypes = vec_extract_wrapped(dtypes);
     Ok(dsl::dtype_cols(dtypes).into())
+}
+
+#[allow(clippy::too_many_arguments)]
+#[pyfunction]
+pub fn duration(
+    days: Option<PyExpr>,
+    seconds: Option<PyExpr>,
+    nanoseconds: Option<PyExpr>,
+    microseconds: Option<PyExpr>,
+    milliseconds: Option<PyExpr>,
+    minutes: Option<PyExpr>,
+    hours: Option<PyExpr>,
+    weeks: Option<PyExpr>,
+) -> PyExpr {
+    set_unwrapped_or_0!(
+        days,
+        seconds,
+        nanoseconds,
+        microseconds,
+        milliseconds,
+        minutes,
+        hours,
+        weeks,
+    );
+    let args = DurationArgs {
+        days,
+        seconds,
+        nanoseconds,
+        microseconds,
+        milliseconds,
+        minutes,
+        hours,
+        weeks,
+    };
+    dsl::duration(args).into()
 }
