+++ py-polars/src/lazy/dsl.rs
@@ -80,20 +80,6 @@ pub fn reduce(lambda: PyObject, exprs: Vec<PyExpr>) -> PyExpr {
     polars::lazy::dsl::reduce_exprs(func, exprs).into()
 }
 
-pub fn cumfold(acc: PyExpr, lambda: PyObject, exprs: Vec<PyExpr>, include_init: bool) -> PyExpr {
-    let exprs = py_exprs_to_exprs(exprs);
-
-    let func = move |a: Series, b: Series| binary_lambda(&lambda, a, b);
-    cumfold_exprs(acc.inner, func, exprs, include_init).into()
-}
-
-pub fn cumreduce(lambda: PyObject, exprs: Vec<PyExpr>) -> PyExpr {
-    let exprs = py_exprs_to_exprs(exprs);
-
-    let func = move |a: Series, b: Series| binary_lambda(&lambda, a, b);
-    cumreduce_exprs(func, exprs).into()
-}
-
 pub fn lit(value: &PyAny, allow_object: bool) -> PyResult<PyExpr> {
     if let Ok(true) = value.is_instance_of::<PyBool>() {
         let val = value.extract::<bool>().unwrap();
